+++ Plantalog/src/plantalog/models/PlantImage.java
@@ -125,6 +125,7 @@ public class PlantImage extends Model {
         return image_id;
     }
     public static void delete(PlantImage p){
-        DBC.execute("DELETE FROM PlantImage WHERE image_id=\"" + p.image_id + "\""); 
+        if(p != null)
+            DBC.execute("DELETE FROM PlantImage WHERE image_id=\"" + p.image_id + "\""); 
     }
 }
