+++ Plantalog/src/plantalog/models/SpecimenRegion.java
@@ -87,6 +87,7 @@ public class SpecimenRegion extends Model {
         DBC.execute("update SpecimenRegion set description = \"" + newDesc + "\" where region_name = \"" + region + "\" ");
     }
     public static void delete(SpecimenRegion r){
-        DBC.execute("DELETE FROM SpecimenRegion WHERE region_id=\"" + r.region_name + "\""); 
+        if(r != null)
+            DBC.execute("DELETE FROM SpecimenRegion WHERE region_id=\"" + r.region_name + "\""); 
     }
 }
